refactor: classify util functions to keep code clean
