--- huaweicloud/util.go
@@ -1,232 +1,0 @@
-package huaweicloud
-
-import (
-	"bytes"
-	"encoding/json"
-	"fmt"
-	"net/http"
-	"sort"
-	"strings"
-	"time"
-
-	"github.com/hashicorp/terraform-plugin-sdk/helper/hashcode"
-	"github.com/hashicorp/terraform-plugin-sdk/helper/resource"
-	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
-	"github.com/huaweicloud/golangsdk"
-	"github.com/unknwon/com"
-)
-
-// BuildRequest takes an opts struct and builds a request body for
-// Gophercloud to execute
-func BuildRequest(opts interface{}, parent string) (map[string]interface{}, error) {
-	b, err := golangsdk.BuildRequestBody(opts, "")
-	if err != nil {
-		return nil, err
-	}
-
-	b = AddValueSpecs(b)
-
-	return map[string]interface{}{parent: b}, nil
-}
-
-// CheckDeleted checks the error to see if it's a 404 (Not Found) and, if so,
-// sets the resource ID to the empty string instead of throwing an error.
-func CheckDeleted(d *schema.ResourceData, err error, msg string) error {
-	if _, ok := err.(golangsdk.ErrDefault404); ok {
-		d.SetId("")
-		return nil
-	}
-
-	return fmt.Errorf("%s: %s", msg, err)
-}
-
-// GetRegion returns the region that was specified in the resource. If a
-// region was not set, the provider-level region is checked. The provider-level
-// region can either be set by the region argument or by HW_REGION_NAME.
-func GetRegion(d *schema.ResourceData, config *Config) string {
-	if v, ok := d.GetOk("region"); ok {
-		return v.(string)
-	}
-
-	return config.Region
-}
-
-func GetEnterpriseProjectID(d *schema.ResourceData, config *Config) string {
-	if v, ok := d.GetOk("enterprise_project_id"); ok {
-		return v.(string)
-	}
-
-	return config.EnterpriseProjectID
-}
-
-// AddValueSpecs expands the 'value_specs' object and removes 'value_specs'
-// from the reqeust body.
-func AddValueSpecs(body map[string]interface{}) map[string]interface{} {
-	if body["value_specs"] != nil {
-		for k, v := range body["value_specs"].(map[string]interface{}) {
-			body[k] = v
-		}
-		delete(body, "value_specs")
-	}
-
-	return body
-}
-
-// MapValueSpecs converts ResourceData into a map
-func MapValueSpecs(d *schema.ResourceData) map[string]string {
-	m := make(map[string]string)
-	for key, val := range d.Get("value_specs").(map[string]interface{}) {
-		m[key] = val.(string)
-	}
-	return m
-}
-
-// MapResourceProp converts ResourceData property into a map
-func MapResourceProp(d *schema.ResourceData, prop string) map[string]interface{} {
-	m := make(map[string]interface{})
-	for key, val := range d.Get(prop).(map[string]interface{}) {
-		m[key] = val.(string)
-	}
-	return m
-}
-
-// List of headers that need to be redacted
-var REDACT_HEADERS = []string{"x-auth-token", "x-auth-key", "x-service-token",
-	"x-storage-token", "x-account-meta-temp-url-key", "x-account-meta-temp-url-key-2",
-	"x-container-meta-temp-url-key", "x-container-meta-temp-url-key-2", "set-cookie",
-	"x-subject-token"}
-
-// RedactHeaders processes a headers object, returning a redacted list
-func RedactHeaders(headers http.Header) (processedHeaders []string) {
-	for name, header := range headers {
-		for _, v := range header {
-			if com.IsSliceContainsStr(REDACT_HEADERS, name) {
-				processedHeaders = append(processedHeaders, fmt.Sprintf("%v: %v", name, "***"))
-			} else {
-				processedHeaders = append(processedHeaders, fmt.Sprintf("%v: %v", name, v))
-			}
-		}
-	}
-	return
-}
-
-// FormatHeaders processes a headers object plus a deliminator, returning a string
-func FormatHeaders(headers http.Header, seperator string) string {
-	redactedHeaders := RedactHeaders(headers)
-	sort.Strings(redactedHeaders)
-
-	return strings.Join(redactedHeaders, seperator)
-}
-
-func checkForRetryableError(err error) *resource.RetryError {
-	switch errCode := err.(type) {
-	case golangsdk.ErrDefault500:
-		return resource.RetryableError(err)
-	case golangsdk.ErrUnexpectedResponseCode:
-		switch errCode.Actual {
-		case 409, 503:
-			return resource.RetryableError(err)
-		default:
-			return resource.NonRetryableError(err)
-		}
-	default:
-		return resource.NonRetryableError(err)
-	}
-}
-
-func suppressEquivilentTimeDiffs(k, old, new string, d *schema.ResourceData) bool {
-	oldTime, err := time.Parse(time.RFC3339, old)
-	if err != nil {
-		return false
-	}
-
-	newTime, err := time.Parse(time.RFC3339, new)
-	if err != nil {
-		return false
-	}
-
-	return oldTime.Equal(newTime)
-}
-
-func validateSubnetV2IPv6Mode(v interface{}, k string) (ws []string, errors []error) {
-	value := v.(string)
-	if value != "slaac" && value != "dhcpv6-stateful" && value != "dhcpv6-stateless" {
-		err := fmt.Errorf("%s must be one of slaac, dhcpv6-stateful or dhcpv6-stateless", k)
-		errors = append(errors, err)
-	}
-	return
-}
-
-func expandToStringSlice(v []interface{}) []string {
-	s := make([]string, len(v))
-	for i, val := range v {
-		if strVal, ok := val.(string); ok {
-			s[i] = strVal
-		}
-	}
-
-	return s
-}
-
-// strSliceContains checks if a given string is contained in a slice
-// When anybody asks why Go needs generics, here you go.
-func strSliceContains(haystack []string, needle string) bool {
-	for _, s := range haystack {
-		if s == needle {
-			return true
-		}
-	}
-	return false
-}
-
-func jsonMarshal(t interface{}) ([]byte, error) {
-	buffer := &bytes.Buffer{}
-	enc := json.NewEncoder(buffer)
-	enc.SetEscapeHTML(false)
-	err := enc.Encode(t)
-	return buffer.Bytes(), err
-}
-
-// Generates a hash for the set hash function used by the ID
-func dataResourceIdHash(ids []string) string {
-	var buf bytes.Buffer
-
-	for _, id := range ids {
-		buf.WriteString(fmt.Sprintf("%s-", id))
-	}
-
-	return fmt.Sprintf("%d", hashcode.String(buf.String()))
-}
-
-// Remove duplicate elements from slice
-func removeDuplicateElem(s []string) []string {
-	result := []string{}
-	tmpMap := map[string]byte{}
-	for _, e := range s {
-		l := len(tmpMap)
-		tmpMap[e] = 0
-		if len(tmpMap) != l {
-			result = append(result, e)
-		}
-	}
-	return result
-}
-
-func removeNil(data map[string]interface{}) map[string]interface{} {
-	withoutNil := make(map[string]interface{})
-
-	for k, v := range data {
-		if v == nil {
-			continue
-		}
-
-		switch v.(type) {
-		case map[string]interface{}:
-			withoutNil[k] = removeNil(v.(map[string]interface{}))
-		default:
-			withoutNil[k] = v
-		}
-	}
-
-	return withoutNil
-}
